Refactor: refatoração da parte de catalogo
