+++ src/main/resources/templates/catalogo.html
@@ -164,6 +164,7 @@
 <div class="container mt-4 mb-4">
 
         <div class="col-md-8">
+
             <div class="search-container">
                 <form th:action="@{/pesquisa}" method="get" class="d-flex flex-grow-1 me-2" style="max-width: 400px;">
                     <input type="text" name="nome" class="form-control"
@@ -191,6 +192,7 @@
                         <i class="fas fa-times"></i>
                     </a>
                 </div>
+
             </div>
         </div>
 </div>
